refactor: SSE response methods separated from handleText
