+++ api/server/utils/index.js
@@ -1,9 +1,11 @@
-const cryptoUtils = require('./crypto');
+const streamResponse = require('./streamResponse');
 const handleText = require('./handleText');
+const cryptoUtils = require('./crypto');
 const citations = require('./citations');
 const sendEmail = require('./sendEmail');
 
 module.exports = {
+  ...streamResponse,
   ...cryptoUtils,
   ...handleText,
   ...citations,
